+++ chalicelib/routes/formDelete.py
@@ -1,11 +1,12 @@
+from chalicelib.models import Form, serialize_model
+from bson.objectid import ObjectId
+from bson import BSON
+
 def form_delete(formId):
-  from ..main import app, TABLES
-  form = TABLES.forms.get_item(
-    Key=dict(id=formId, version=1),
-    ProjectionExpression="cff_permissions"
-  )["Item"]
-  app.check_permissions(form, 'Forms_Delete')
-  TABLES.forms.delete_item(
-    Key=dict(id=formId, version=1)
-  )
-  return {"res": None, "success": True, "action": "delete"}
+    """Creates a new form with a blank schema and uiSchema.
+    """
+    from ..main import app
+    form = Form.objects.only("cff_permissions").get({"_id": ObjectId(formId)})
+    app.check_permissions(form, 'Forms_Delete')
+    form.delete()
+    return {"res": None, "success": True, "action": "delete", "formId": formId}