+++ chalicelib/routes/formPermissions.py
@@ -1,6 +1,12 @@
-from boto3.dynamodb.conditions import Key
+from chalicelib.models import Form, serialize_model
+from bson.objectid import ObjectId
 from pydash.arrays import union
 POSSIBLE_PERMISSIONS = ["owner", "Responses_View", "Responses_Export", "Responses_ViewSummary", "Responses_Edit", "Responses_CheckIn", "Forms_Edit", "Forms_PermissionsView", "Forms_PermissionsEdit"] # "Form_PermissionsView", "Form_PermissionsEdit", "Forms_List", "Schemas_List", "SchemaModifiers_Edit"]
+
+def form_render(formId):
+    """Get forms user has access to."""
+    form = Form.objects.get(id=ObjectId(formId)).only("name", "date_created", "date_last_modified", "schema", "uiSchema", "formOptions")
+    return {"res": form}
 
 def form_get_permissions(formId):
   """
@@ -9,27 +15,15 @@ def form_get_permissions(formId):
   query_params:
   ?mine=1 -- only get my permission names (for logged in user) - requires no auth except for being logged in. (This is currently not used).
   """
-  from ..main import app, TABLES, dynamodb, get_table_name
-  form = TABLES.forms.get_item(
-      Key=dict(id=formId, version=1),
-      ProjectionExpression="cff_permissions"
-  )["Item"]
+  from ..main import app
+  form = Form.objects.get(id=ObjectId(formId)).only("cff_permissions")
   if app.current_request.query_params and "mine" in app.current_request.query_params:
     permissions = app.get_user_permissions(app.get_current_user_id(), form)
     return {"res": {"permissions": permissions}}
   app.check_permissions(form, 'Forms_PermissionsView')
-  userIds = form['cff_permissions'].keys()
-  # print([{"id": userId} for userId in userIds])
-  users = dynamodb.batch_get_item(RequestItems={
-    get_table_name("users"): {
-      "Keys": [{"id": userId} for userId in userIds],
-      "ProjectionExpression": "#name, email, id",
-      "ExpressionAttributeNames": {"#name": "name"},
-      "ConsistentRead": True
-    }
-  })["Responses"][get_table_name("users")]
-  user_lookup = {user["id"]: user for user in users}
-  return {"res": {"permissions": form['cff_permissions'], "userLookup": user_lookup, "possiblePermissions": POSSIBLE_PERMISSIONS}}
+  userIds = form.cff_permissions.keys()
+  user_lookup = {userId: {"name": "Test", "email": "test@test.com", "id": userId} for userId in userIds}
+  return {"res": {"permissions": form.cff_permissions, "userLookup": user_lookup, "possiblePermissions": POSSIBLE_PERMISSIONS}}
 
 def form_edit_permissions(formId):
   """Set form permissions of a particular user to an array.
@@ -39,11 +33,8 @@ def form_edit_permissions(formId):
     "permissions": ["Responses_Edit", "Responses_View", ""] or string. -- should have all permissions you want user to be assigned to.
   }
   """
-  from ..main import app, TABLES
-  form = TABLES.forms.get_item(
-      Key=dict(id=formId, version=1),
-      ProjectionExpression="cff_permissions"
-  )["Item"]
+  from ..main import app
+  form = Form.objects.get(id=ObjectId(formId)).only("cff_permissions")
   app.check_permissions(form, 'Forms_PermissionsEdit')
   permissions = app.current_request.json_body['permissions']
   userId = app.current_request.json_body['userId']
@@ -55,12 +46,6 @@ def form_edit_permissions(formId):
     if type(v) is not bool:
       raise Exception("Permission {} not formatted correctly; each value should be a boolean.".format({i: v}))
   # todo: update date last modified, here, too?
-  updated_form = TABLES.forms.update_item(
-    Key=dict(id=formId, version=1),
-    UpdateExpression="SET cff_permissions.#uid = :permissions ",
-    ExpressionAttributeValues={":permissions": permissions },
-    ExpressionAttributeNames={"#uid": userId},
-    ReturnValues="ALL_NEW"
-  )["Attributes"]
-  return {"res": updated_form['cff_permissions'], "success": True, "action": "update"}
+  form.update(**{f"cff_permissions.{userId}": permissions})
+  return {"res": serialize_model(Form)["cff_permissions"], "success": True, "action": "update"}