+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Todo: more rigorous tests for response checkin endpoint, also for view responses for people with only Responses_CheckIn permission.
 Todo: version api urls (/1.x/...)
 
+# 2.0.0 (tbd)
+- Add mongodb (cosmosdb)
+- Use pymodm
+- Endpoint changes to be more RESTful.
+
+/forms/{formId}/render -> /forms/{formId}
+/centers/{centerId}/forms - removed
+/centers/{centerId}/schemas - removed
+/centers - removed
+
 # 1.1.14 (5/26/18)
 - Fix responseSendConfirmationEmail endpoint to work with v2 version of form
 - Allow undefined variables to not crash jinja templates in email; add format_payment filter to templates too.